+++ src/no/northcode/jens/intranetsek2tg/TimeListener.java
@@ -2,19 +2,11 @@ package no.northcode.jens.intranetsek2tg;
 
 import java.time.LocalDate;
 import java.util.HashMap;
-import java.util.List;
-
 import org.quartz.Job;
-import org.quartz.JobDetail;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
-import org.quartz.JobKey;
 import org.quartz.SchedulerContext;
 import org.quartz.SchedulerException;
-import org.quartz.SchedulerListener;
-import org.quartz.Trigger;
-import org.quartz.TriggerKey;
-
 import pro.zackpollard.telegrambot.api.TelegramBot;
 import pro.zackpollard.telegrambot.api.chat.Chat;
 
@@ -35,7 +27,7 @@ public class TimeListener implements Job {
         TelegramListener listener = (TelegramListener) schedulerContext.get("telegramListener");
         
         HashMap<String, GroupData> groups = listener.getGroups();
-        HashMap<Integer, UserData> users = listener.getUsers();
+        HashMap<Long, UserData> users = listener.getUsers();
         
         
 		for(String key : groups.keySet()) {
